moving classes to a more component like structure
- add Wikipedia and YouTube namespace and moved applicable classes.
  this way the code is clustered by implementation making it easier
  to copy paste it to your own application.
