+++ src/Rag/Blog/Loader.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Blog;
+namespace App\Rag\Blog;
 
 use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\Uid\Uuid;
