+++ tests/Rag/Blog/PostTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Blog;
+namespace App\Tests\Rag\Blog;
 
-use App\Blog\Post;
+use App\Rag\Blog\Post;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Uid\Uuid;
